some reorganizing
